% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplatemap.R
\name{add_single_field}
\alias{add_single_field}
\title{add_single_field}
\usage{
add_single_field(
  pm,
  tab_name,
  id,
  name,
  required = FALSE,
  type = c("text", "numeric", "boolean", "select", "multiselect"),
  placeholder = "",
  options = NULL,
  units = NULL,
  defaultUnit = NULL
)
}
\description{
create a single field
}
