% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplatemap.R
\name{rplatemap}
\alias{rplatemap}
\title{rplatemap}
\usage{
rplatemap(
  nrow = 8,
  ncol = 12,
  readOnly = FALSE,
  attributes = NULL,
  data = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
R binding to plate-map javascript library
}
